Load more things with get_classes. Remove '_' prefix from some files names, as they're not just used internally. Rename inheritable resources to same odin name, as this is just what should be used for oscar resources if you want them to be able to be subclassed.
