+++ oscar_odin/resources/__init__.py
@@ -2,6 +2,6 @@
 
 from . import address, auth, catalogue, order
 
-from ._base import OscarResource
+from .base import OscarResource
 
 __all__ = ["OscarResource", "address", "catalogue", "order", "auth"]
